Extract namespace lookup optimisation into separate pass
