+++ test/function/namespace-optimisation-before-exports/_config.js
@@ -1,0 +1,5 @@
+module.exports = {
+	description: 'namespace optimisation must be done after all exports are defined'
+};
+
+// See: https://github.com/rollup/rollup/issues/148
