+++ test/function/namespace-optimisation-before-exports/bar.js
@@ -1,0 +1,1 @@
+export default function bar () {}
