+++ test/function/namespace-optimisation-before-exports/foo.js
@@ -1,0 +1,1 @@
+export { default as bar } from './bar.js';
