+++ test/function/namespace-optimisation-before-exports/main.js
@@ -1,0 +1,6 @@
+import * as foo from './foo';
+import './zoo';
+
+export default {
+	foo: foo
+};
