+++ test/function/namespace-optimisation-before-exports/zoo.js
@@ -1,0 +1,7 @@
+import * as foo from './foo';
+
+function wah () {
+	foo.bar();
+}
+
+wah();
